requirements.txt
# Core Dependencies
# Python 3.11

# Frontend Framework
streamlit==1.28.1
streamlit-option-menu==0.3.2

# Backend
fastapi==0.104.1
uvicorn==0.24.0
pydantic==1.10.13

# LLM & NLP
sentence-transformers==2.2.2
transformers==4.34.1
# Note: huggingface-hub is automatically installed by sentence-transformers and transformers
# Do NOT pin it manually - let pip resolve dependencies

# Vector Databases
faiss-cpu==1.7.4
hnswlib==0.8.0

# Machine Learning & Data Science
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.26.3
scipy==1.11.4

# Data Processing & Validation
python-dotenv==1.0.0
requests==2.31.0
aiohttp==3.9.1

# Utilities & Logging
tqdm==4.66.1
joblib==1.3.2

# Retry Logic
tenacity==8.2.3

# OpenAI Integration (requires >= 1.10.0 for OpenAI client)
openai==1.12.0

# Note: setuptools and wheel are managed by pip automatically
# Do NOT pin them manually on Windows - it can cause build issues
